update js to tsx for navbar and footer
