+++ client/src/App.js
@@ -4,12 +4,11 @@ import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'; //use
 import { ApolloClient, InMemoryCache, ApolloProvider } from '@apollo/client';
 
 import Home from './pages/Home';
-// import About from './pages/About';
 // import Contact from './pages/Contact';
 //import Search from './pages/SearchBook';
-// import NavMenu from './components/NavMenu';
-import Footer from './components/Footer';
-import Navbar from './components/NavBar';
+import NavMenu from './components/NavMenu';
+import Footer from './components/Footer.tsx';
+import Navbar from './components/NavBar.tsx';
 
 // Extend the theme to include custom colors, fonts, etc
 // colorname.900
