+++ client/src/components/HomeDetails.js
@@ -1,5 +1,6 @@
 import * as React from "react";
 import { useState } from 'react';
+
 import openBookPic from '../images/photos/olga-tutunaru-JMATuFkXeHU-unsplash.jpg';
 import {
   Flex,
